+++ tests/trace_compiled_dfa_as_dot.rs
@@ -1,10 +1,13 @@
 // Outputs the compiled automata as in dot format for all the modes files in the data directory.
 // Run with `cargo test -- --nocapture trace_compiled_dfa_as_dot`
 
+#[cfg(not(feature = "regex_automata"))]
 use std::fs;
 
+#[cfg(not(feature = "regex_automata"))]
 use scnr::{ScannerBuilder, ScannerMode};
 
+#[cfg(not(feature = "regex_automata"))]
 #[test]
 fn trace_compiled_dfa_as_dot() {
     // Initialize the logger